+++ src/components/Input/DynamicContent/ContentValueField/ContentValueField.spec.tsx
@@ -1,5 +1,6 @@
 import { test, expect } from '../../../../../playwright/ct-test';
-import ContentValueFieldTestWrapper, { buildDescriptor } from './ContentValueFieldTestWrapper';
+import ContentValueFieldTestWrapper from './ContentValueFieldTestWrapper';
+import { buildDescriptor } from './ContentValueFieldTestDescriptors';
 import { AttributeContentType } from 'types/openapi';
 
 type Page = import('@playwright/test').Page;
