+++ src/components/Input/DynamicContent/ContentValueField/ContentValueFieldTestWrapper.tsx
@@ -2,17 +2,6 @@ import React from 'react';
 import * as ReactHookForm from 'react-hook-form';
 import ContentValueField from './index';
 import type { BaseAttributeContentModel, CustomAttributeModel } from 'types/attributes';
-import { AttributeContentType, AttributeType } from 'types/openapi';
-
-const defaultProperties: CustomAttributeModel['properties'] = {
-    label: 'Test',
-    visible: true,
-    required: false,
-    readOnly: false,
-    list: false,
-    multiSelect: false,
-    extensibleList: false,
-};
 
 export type ContentValueFieldTestWrapperProps = Readonly<{
     id?: string;
@@ -20,22 +9,6 @@ export type ContentValueFieldTestWrapperProps = Readonly<{
     initialContent?: BaseAttributeContentModel[];
     onSubmit?: (attributeUuid: string, content: BaseAttributeContentModel[]) => void;
 }>;
-
-export function buildDescriptor(overrides: Partial<CustomAttributeModel> = {}): CustomAttributeModel {
-    const base: CustomAttributeModel = {
-        uuid: 'test-uuid',
-        name: 'testAttr',
-        type: AttributeType.Custom,
-        contentType: AttributeContentType.String,
-        content: [],
-        properties: defaultProperties,
-    };
-    return {
-        ...base,
-        ...overrides,
-        properties: { ...defaultProperties, ...overrides.properties },
-    };
-}
 
 function ContentValueFieldTestWrapper({ id, descriptor, initialContent, onSubmit = () => {} }: ContentValueFieldTestWrapperProps) {
     const methods = ReactHookForm.useForm({
